+++ packages/algoliasearch/index.js
@@ -1,2 +1,15 @@
-// eslint-disable-next-line functional/immutable-data, import/no-commonjs
-module.exports = require('./dist/algoliasearch.cjs.js');
+/* eslint-disable functional/immutable-data, import/no-commonjs */
+const algoliasearch = require('./dist/algoliasearch.cjs.js');
+
+/**
+ * The Common JS build is the default entry point for the Node environment. Keep in
+ * in mind, that for the browser environment, we hint the bundler to use the UMD
+ * build instead as specified on the key `browser` of our `package.json` file.
+ */
+module.exports = algoliasearch;
+
+/**
+ * In addition, we also set explicitly the default export below making
+ * this Common JS module in compliance with es6 modules specification.
+ */
+module.exports.default = algoliasearch;