+++ packages/client-search/src/__tests__/integration/batching.test.ts
@@ -3,8 +3,6 @@ import { ObjectWithObjectID } from '../../types';
 
 const testSuite = new TestSuite('batching');
 
-afterAll(() => testSuite.cleanUp());
-
 test(testSuite.testName, async () => {
   const index = testSuite.makeIndex();