+++ CHANGELOG.md
@@ -6,7 +6,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # Release Notes
 
-## [Unreleased](https://github.com/algolia/algoliasearch-client-javascript/compare/4.0.1...master)
+## [Unreleased](https://github.com/algolia/algoliasearch-client-javascript/compare/4.0.3...master)
+
+### Fixed
+- Field `abtests` in GetAbTestsResponse ([#1012](https://github.com/algolia/algoliasearch-client-javascript/pull/1012))
+
+## [v4.0.3](https://github.com/algolia/algoliasearch-client-javascript/compare/4.0.2...4.0.3)
+
+### Fixed
+- `browseObjects` missing cursor error ([#1011](https://github.com/algolia/algoliasearch-client-javascript/pull/1011))
+- `browseSynonyms|browseRules` giving just a maximum of 20 hits ([#1011](https://github.com/algolia/algoliasearch-client-javascript/pull/1011))
+- Node entry point not es6 compliant ([#1005](https://github.com/algolia/algoliasearch-client-javascript/pull/1005))
+
+## [v4.0.2](https://github.com/algolia/algoliasearch-client-javascript/compare/4.0.1...4.0.2)
+
+### Fixed
+- DOMException being thrown while instantiating browser local storage cache driver ([#999](https://github.com/algolia/algoliasearch-client-javascript/pull/999))
 
 ## [v4.0.1](https://github.com/algolia/algoliasearch-client-javascript/compare/4.0.0...4.0.1)